Added ArrayedList with RemoveIfs.... Committed all
