+++ src/main/java/appiumPackage/ConfigurationReader.java
@@ -1,0 +1,26 @@
+package appiumPackage;
+
+import java.io.FileInputStream;
+import java.io.IOException;
+import java.io.InputStream;
+import java.util.Properties;
+
+public class ConfigurationReader {
+
+    private static Properties properties;
+
+    static {
+        try (InputStream inputStream = new FileInputStream("configuration.properties")) {
+            properties = new Properties();
+            properties.load(inputStream);
+        } catch (IOException e) {
+            e.printStackTrace();
+            throw new RuntimeException("Failed to load properties file!");
+        }
+    }
+
+    public static String getProperty(String property) {
+        return properties.getProperty(property);
+    }
+}
+
