+++ src/main/java/appiumPackage/Driver.java
@@ -1,0 +1,76 @@
+package appiumPackage;
+
+import io.appium.java_client.AppiumDriver;
+import io.appium.java_client.MobileElement;
+import io.appium.java_client.android.AndroidDriver;
+import io.appium.java_client.remote.MobileCapabilityType;
+import org.openqa.selenium.Platform;
+import org.openqa.selenium.remote.DesiredCapabilities;
+
+import java.net.MalformedURLException;
+import java.net.URL;
+import java.util.Objects;
+
+public class Driver {
+
+    private static AppiumDriver<MobileElement> driver;
+    private static URL url;
+
+    private Driver(){
+
+    }
+
+    public static AppiumDriver<MobileElement> getDriver() {
+        String platform = ConfigurationReader.getProperty("platform");
+        if (Objects.isNull(driver)) {
+            switch (platform) {
+                case "android-calculator":
+                    DesiredCapabilities desiredCapabilities = new DesiredCapabilities();
+                    desiredCapabilities.setCapability(MobileCapabilityType.AUTOMATION_NAME, "UiAutomator2");
+                    desiredCapabilities.setCapability(MobileCapabilityType.PLATFORM_NAME, Platform.ANDROID);
+                    desiredCapabilities.setCapability(MobileCapabilityType.PLATFORM_VERSION, "10.0");
+                    desiredCapabilities.setCapability(MobileCapabilityType.DEVICE_NAME, "Pixel 3");
+                    desiredCapabilities.setCapability(MobileCapabilityType.APP, "https://cybertek-appium.s3.amazonaws.com/calculator.apk");
+                    try {
+                        url = new URL("http://localhost:4723/wd/hub");
+                    } catch (MalformedURLException e) {
+                        e.printStackTrace();
+                    }
+                    driver = new AndroidDriver<>(url, desiredCapabilities);
+                    break;
+
+
+                case "android-swaglab":
+                    DesiredCapabilities desiredCapabilities2 = new DesiredCapabilities();
+                    desiredCapabilities2.setCapability(MobileCapabilityType.AUTOMATION_NAME, "UiAutomator2");
+                    desiredCapabilities2.setCapability(MobileCapabilityType.PLATFORM_NAME, Platform.ANDROID);
+                    desiredCapabilities2.setCapability(MobileCapabilityType.PLATFORM_VERSION, "10.0");
+                    desiredCapabilities2.setCapability(MobileCapabilityType.DEVICE_NAME, "Pixel 3");
+                    desiredCapabilities2.setCapability(MobileCapabilityType.APP, "C:\\Users\\Saron Tsegay\\IdeaProjects\\AppiumTestsB26\\Android.SauceLabs.Mobile.Sample.app.2.7.1.apk");
+
+
+                    desiredCapabilities2.setCapability("appPackage", "com.swaglabsmobileapp");
+                    desiredCapabilities2.setCapability("appActivity", "com.swaglabsmobileapp.SplashActivity");
+
+
+
+                    try {
+                        url = new URL("http://localhost:4723/wd/hub");
+                    } catch (MalformedURLException e) {
+                        e.printStackTrace();
+                    }
+                    driver = new AndroidDriver<>(url, desiredCapabilities2);
+                    break;
+
+            }
+        }
+        return driver;
+    }
+
+    public static void closeDriver() {
+        if (Objects.nonNull(driver)) {
+            driver.closeApp();
+            driver = null;
+        }
+    }
+}
