+++ src/main/java/appiumPackage/WebTestBaser.java
@@ -1,0 +1,41 @@
+package appiumPackage;
+
+import io.appium.java_client.AppiumDriver;
+import io.appium.java_client.MobileElement;
+import io.appium.java_client.android.AndroidDriver;
+import io.appium.java_client.remote.MobileCapabilityType;
+//import org.junit.jupiter.api.AfterEach;
+//import org.junit.jupiter.api.BeforeEach;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.remote.DesiredCapabilities;
+import org.testng.annotations.Test;
+
+import java.net.MalformedURLException;
+import java.net.URL;
+
+public abstract class WebTestBaser {
+
+    protected WebDriver driver;
+
+    @Test
+    public void initialTest() throws MalformedURLException {
+
+        DesiredCapabilities caps = new DesiredCapabilities();
+
+        caps.setCapability(MobileCapabilityType.DEVICE_NAME, "Pixel 3");
+        caps.setCapability(MobileCapabilityType.PLATFORM_NAME, "Android");
+        caps.setCapability(MobileCapabilityType.PLATFORM_VERSION, "10.0");
+        caps.setCapability(MobileCapabilityType.AUTOMATION_NAME, "UiAutomator2");
+        caps.setCapability(MobileCapabilityType.BROWSER_NAME, "chrome");
+        caps.setCapability(MobileCapabilityType.APP, "https://cybertek-appium.s3.amazonaws.com/calculator.apk");
+
+        URL url = new URL("http://localhost:4723/wd/hub");
+
+        driver = new AndroidDriver<MobileElement>(url, caps);
+
+
+    }
+
+//        driver.close();
+    }
+
